Update result display format

+++ src/main/java/com/ssmtariq/srlab/jtanalyzer/GenericTree.java
@@ -132,9 +132,9 @@ public class GenericTree {
     public static void aggregateResult(Map<String, Double> serviceDuration) {
         serviceDuration.forEach((k,v)->{
             if (aggregatorMap.containsKey(k)) {
-                aggregatorMap.put(k, (v.longValue()+aggregatorMap.get(k)));
+                aggregatorMap.put(k, (v.intValue()+aggregatorMap.get(k)));
             } else {
-                aggregatorMap.put(k, v.longValue());
+                aggregatorMap.put(k, (long) v.intValue());
             }
         });
     }
